Migrate to Click cli commands
